refactor(app): extract useGoToLabel hook (95j)
Moves goToLabelItems, goToLabelCanvasName, warpLabelItems, and handleGoToLabel
out of App.tsx into src/hooks/useGoToLabel.ts. Pure memo + callback — no state,
no side effects.
