+++ src/App.tsx
@@ -61,6 +61,7 @@ import { useExternalFileChanges } from '@/hooks/useExternalFileChanges';
 import { useGameExecution } from '@/hooks/useGameExecution';
 import { useWarpLaunch } from '@/hooks/useWarpLaunch';
 import { useMenuCommandDispatch } from '@/hooks/useMenuCommandDispatch';
+import { useGoToLabel } from '@/hooks/useGoToLabel';
 import { formatErrorMessage } from '@/lib/formatErrorMessage';
 import { computeRouteCanvasLayout } from '@/lib/routeCanvasLayout';
 import { resolveWarpTarget } from '@/lib/warpTarget';
@@ -1211,38 +1212,15 @@ const App: React.FC = () => {
   const activeCanvasTabId = activeTabId === 'canvas' || activeTabId === 'route-canvas' || activeTabId === 'choice-canvas'
     ? activeTabId : null;
 
-  const goToLabelItems = useMemo<GoToLabelItem[]>(() => {
-    if (activeCanvasTabId === 'canvas') {
-      return analysisResult.labelNodes.map(n => ({ label: n.label, id: n.blockId }));
-    }
-    if (activeCanvasTabId === 'route-canvas' || activeCanvasTabId === 'choice-canvas') {
-      return routeAnalysisResult.labelNodes.map(n => ({ label: n.label, id: n.id }));
-    }
-    return [];
-  }, [activeCanvasTabId, analysisResult.labelNodes, routeAnalysisResult.labelNodes]);
-
-  const goToLabelCanvasName = activeCanvasTabId === 'canvas' ? 'Story'
-    : activeCanvasTabId === 'route-canvas' ? 'Route'
-    : activeCanvasTabId === 'choice-canvas' ? 'Choice'
-    : '';
-
-  const warpLabelItems = useMemo<GoToLabelItem[]>(() => {
-    return Object.values(analysisResult.labels)
-      .slice()
-      .sort((a, b) => a.label.localeCompare(b.label))
-      .map(loc => ({ label: loc.label, id: loc.label }));
-  }, [analysisResult.labels]);
-
-  const handleGoToLabel = useCallback((id: string) => {
-    closeGoToLabelModal();
-    if (activeCanvasTabId === 'canvas') {
-      setCenterOnBlockRequest({ blockId: id, key: Date.now() });
-    } else if (activeCanvasTabId === 'route-canvas') {
-      setCenterOnRouteNodeRequest({ nodeId: id, key: Date.now() });
-    } else if (activeCanvasTabId === 'choice-canvas') {
-      setCenterOnChoiceNodeRequest({ nodeId: id, key: Date.now() });
-    }
-  }, [activeCanvasTabId, closeGoToLabelModal, setCenterOnBlockRequest, setCenterOnChoiceNodeRequest, setCenterOnRouteNodeRequest]);
+  const { goToLabelItems, goToLabelCanvasName, warpLabelItems, handleGoToLabel } = useGoToLabel({
+    activeCanvasTabId,
+    analysisResult,
+    routeAnalysisResult,
+    closeGoToLabelModal,
+    setCenterOnBlockRequest,
+    setCenterOnRouteNodeRequest,
+    setCenterOnChoiceNodeRequest,
+  });
 
   const {
     isGameRunning,
